backend/requirements.txt
asgiref==3.11.0
Django==5.0
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.11.0
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.5
tzdata==2025.3
